@@ -1,5 +1,7 @@
 ﻿using FluentAssertions;
+using MediatR;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.Extensions.Logging;
 using Moq;
 using NUnit.Framework;
 using SFA.DAS.Apprenticeships.Api.Controllers;
@@ -17,16 +19,17 @@ namespace SFA.DAS.Apprenticeships.Api.UnitTests.Controllers.Apprenticeship
         [Test, MoqAutoData]
         public async Task Then_Gets_ApprenticeshipKey_From_ApiClient(
             Guid expectedResponse,
-            Mock<IApprenticeshipsApiClient<ApprenticeshipsApiConfiguration>> mockApprenticeshipsApiClient,
-            Mock<ICommitmentsV2ApiClient<CommitmentsV2ApiConfiguration>> mockCommitmentsV2ApiApiClient)
+			Mock<ILogger<ApprenticeshipController>> mockLogger,
+			Mock<IMediator> mockMediator,
+			Mock<IApprenticeshipsApiClient<ApprenticeshipsApiConfiguration>> mockApprenticeshipsApiClient)
         {
             //  Arrange
             mockApprenticeshipsApiClient.Setup(x => x.Get<Guid>(It.IsAny<GetApprenticeshipKeyRequest>()))
                 .ReturnsAsync(expectedResponse);
-            var controller = new ApprenticeshipController(mockApprenticeshipsApiClient.Object, mockCommitmentsV2ApiApiClient.Object);
+			var controller = new ApprenticeshipController(mockLogger.Object, mockApprenticeshipsApiClient.Object, Mock.Of<ICommitmentsV2ApiClient<CommitmentsV2ApiConfiguration>>(), mockMediator.Object);
 
-            //  Act
-            var result = await controller.GetApprenticeshipKey("anyApprenticeshipHashedId");
+			//  Act
+			var result = await controller.GetApprenticeshipKey("anyApprenticeshipHashedId");
 
             //  Assert
             var okObjectResult = result.ShouldBeOfType<OkObjectResult>();
