@@ -1,13 +1,13 @@
 ﻿using FluentAssertions;
+using MediatR;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.Extensions.Logging;
 using Moq;
 using NUnit.Framework;
 using SFA.DAS.Apprenticeships.Api.Controllers;
-using SFA.DAS.Apprenticeships.Api.Models;
-using SFA.DAS.Apprenticeships.InnerApi;
+using SFA.DAS.Apprenticeships.Application.Apprenticeship;
+using SFA.DAS.Apprenticeships.Responses;
 using SFA.DAS.SharedOuterApi.Configuration;
-using SFA.DAS.SharedOuterApi.InnerApi.Requests.Apprenticeships;
-using SFA.DAS.SharedOuterApi.InnerApi.Responses.Apprenticeships;
 using SFA.DAS.SharedOuterApi.Interfaces;
 using SFA.DAS.Testing.AutoFixture;
 using System;
@@ -15,41 +15,26 @@ using System.Threading.Tasks;
 
 namespace SFA.DAS.Apprenticeships.Api.UnitTests.Controllers.Apprenticeship
 {
-    public class WhenGettingApprenticeshipPrice
+	public class WhenGettingApprenticeshipPrice
     {
         [Test, MoqAutoData]
         public async Task Then_Gets_ApprenticeshipPrice_From_ApiClient(
             ApprenticeshipPriceResponse expectedResponse,
-            Mock<IApprenticeshipsApiClient<ApprenticeshipsApiConfiguration>> mockApprenticeshipsApiClient,
-            Mock<ICommitmentsV2ApiClient<CommitmentsV2ApiConfiguration>> mockCommitmentsV2ApiApiClient)
+			Mock<ILogger<ApprenticeshipController>> mockLogger,
+			Mock<IMediator> mockMediator,
+			Mock<IApprenticeshipsApiClient<ApprenticeshipsApiConfiguration>> mockApprenticeshipsApiClient)
         {
             //  Arrange
-            mockApprenticeshipsApiClient.Setup(x=>x.Get<GetApprenticeshipPriceResponse>(It.IsAny<GetApprenticeshipPriceRequest>()))
-                .ReturnsAsync(new GetApprenticeshipPriceResponse
-                {
-                    AccountLegalEntityId = 1,
-                    ApprenticeshipKey = expectedResponse.ApprenticeshipKey,
-                    ApprenticeshipActualStartDate = expectedResponse.ApprenticeshipActualStartDate,
-                    ApprenticeshipPlannedEndDate = expectedResponse.ApprenticeshipPlannedEndDate,
-                    AssessmentPrice = expectedResponse.AssessmentPrice,
-                    EarliestEffectiveDate = expectedResponse.EarliestEffectiveDate,
-                    FundingBandMaximum = expectedResponse.FundingBandMaximum,
-                    TrainingPrice = expectedResponse.TrainingPrice
-                });
-
-            mockCommitmentsV2ApiApiClient.Setup(x => x.Get<GetAccountLegalEntityResponse>(It.IsAny<GetAccountLegalEntityRequest>()))
-                .ReturnsAsync(new GetAccountLegalEntityResponse
-                {
-                    LegalEntityName = expectedResponse.EmployerName!
-                });
+            mockMediator.Setup(x => x.Send(It.IsAny<GetApprenticeshipPriceQuery>(), default)).ReturnsAsync(expectedResponse);
 
             mockCommitmentsV2ApiApiClient.Setup(x => x.Get<GetProviderResponse>(It.IsAny<GetProviderRequest>()))
                 .ReturnsAsync(new GetProviderResponse
                 {
                     Name = expectedResponse.ProviderName
                 });
+                
+                var controller = new ApprenticeshipController(mockLogger.Object, mockApprenticeshipsApiClient.Object, Mock.Of<ICommitmentsV2ApiClient<CommitmentsV2ApiConfiguration>>());
 
-            var controller = new ApprenticeshipController(mockApprenticeshipsApiClient.Object, mockCommitmentsV2ApiApiClient.Object);
 
             //  Act
             var result = await controller.GetApprenticeshipPrice(Guid.NewGuid());
