@@ -1,5 +1,7 @@
-﻿using Microsoft.AspNetCore.Mvc;
+﻿using MediatR;
+using Microsoft.AspNetCore.Mvc;
 using SFA.DAS.Apprenticeships.Api.Models;
+using SFA.DAS.Apprenticeships.Application.Apprenticeship;
 using SFA.DAS.Apprenticeships.InnerApi;
 using SFA.DAS.SharedOuterApi.Configuration;
 using SFA.DAS.SharedOuterApi.InnerApi.Requests.Apprenticeships;
@@ -15,47 +17,47 @@ namespace SFA.DAS.Apprenticeships.Api.Controllers
     {
         private readonly IApprenticeshipsApiClient<ApprenticeshipsApiConfiguration> _apiClient;
         private readonly ICommitmentsV2ApiClient<CommitmentsV2ApiConfiguration> _apiCommitmentsClient;
+        private readonly IMediator _mediator;
+        private readonly ILogger<ApprenticeshipController> _logger;
 
-        public ApprenticeshipController(IApprenticeshipsApiClient<ApprenticeshipsApiConfiguration> apiClient, ICommitmentsV2ApiClient<CommitmentsV2ApiConfiguration> apiCommitmentsClient)
+		public ApprenticeshipController(
+			ILogger<ApprenticeshipController> logger,
+			IApprenticeshipsApiClient<ApprenticeshipsApiConfiguration> apiClient,
+            ICommitmentsV2ApiClient<CommitmentsV2ApiConfiguration> apiCommitmentsClient,
+            IMediator mediator)
         {
+            _logger = logger;
             _apiClient = apiClient;
             _apiCommitmentsClient = apiCommitmentsClient;
-        }
+            _mediator = mediator;
+		}
 
         [HttpGet]
         [Route("{apprenticeshipKey}/price")]
         public async Task<ActionResult> GetApprenticeshipPrice(Guid apprenticeshipKey)
         {
-            var apprenticePriceInnerModel = await _apiClient.Get<GetApprenticeshipPriceResponse>(new GetApprenticeshipPriceRequest { ApprenticeshipKey = apprenticeshipKey });
-            
-            if(apprenticePriceInnerModel == null)
-            {
-                return NotFound();
-            }
+			try
+			{
+				var apprenticeshipPriceResponse = await _mediator.Send(new GetApprenticeshipPriceQuery(apprenticeshipKey));
 
-            string? employerName = null;
+				if (apprenticeshipPriceResponse == null)
+				{
+					return NotFound();
+				}
+
+				return Ok(apprenticeshipPriceResponse);
             if(apprenticePriceInnerModel.AccountLegalEntityId.HasValue)
-            {
-                var employer = await _apiCommitmentsClient.Get<GetAccountLegalEntityResponse>(new GetAccountLegalEntityRequest(apprenticePriceInnerModel.AccountLegalEntityId.Value));
-                employerName = employer?.LegalEntityName;
-            }
+			}
 
+			catch (Exception e)
             var provider = await _apiCommitmentsClient.Get<GetProviderResponse>(new GetProviderRequest(apprenticePriceInnerModel.UKPRN));
             
-            var apprenticeshipPriceOuterModel = new ApprenticeshipPriceResponse
-            {
-                ApprenticeshipKey = apprenticePriceInnerModel!.ApprenticeshipKey,
-                ApprenticeshipActualStartDate = apprenticePriceInnerModel.ApprenticeshipActualStartDate,
-                ApprenticeshipPlannedEndDate = apprenticePriceInnerModel.ApprenticeshipPlannedEndDate,
-                AssessmentPrice = apprenticePriceInnerModel.AssessmentPrice,
-                EarliestEffectiveDate = apprenticePriceInnerModel.EarliestEffectiveDate,
-                FundingBandMaximum = apprenticePriceInnerModel.FundingBandMaximum,
-                TrainingPrice = apprenticePriceInnerModel.TrainingPrice,
+			{
+				_logger.LogError(e, "Error attempting to get ApprenticeshipPrice");
+				return BadRequest();
+			}
                 EmployerName = employerName,
                 ProviderName = provider.Name
-            };
-
-            return Ok(apprenticeshipPriceOuterModel);
         }
 
         [HttpGet]
@@ -70,26 +72,47 @@ namespace SFA.DAS.Apprenticeships.Api.Controllers
         public async Task<ActionResult> CreateApprenticeshipPriceChange(Guid apprenticeshipKey,
             [FromBody] CreateApprenticeshipPriceChangeRequest request)
         {
-            await _apiClient.PostWithResponseCode<object>(new PostCreateApprenticeshipPriceChangeRequest(
-                apprenticeshipKey,
-                request.ProviderId,
-                request.EmployerId,
-                request.UserId,
-                request.TrainingPrice,
-                request.AssessmentPrice,
-                request.TotalPrice,
-                request.Reason,
-                request.EffectiveFromDate
-            ), false);
-            return Ok();
-        }
+			var response = await _apiClient.PostWithResponseCode<object>(new PostCreateApprenticeshipPriceChangeRequest(
+				apprenticeshipKey,
+				request.ProviderId,
+				request.EmployerId,
+				request.UserId,
+				request.TrainingPrice,
+				request.AssessmentPrice,
+				request.TotalPrice,
+				request.Reason,
+				request.EffectiveFromDate
+			), false);
+
+			if (!string.IsNullOrEmpty(response.ErrorContent))
+			{
+				return BadRequest();
+			}
+			return Ok();
+		}
 
         [HttpGet]
         [Route("{apprenticeshipKey}/priceHistory/pending")]
         public async Task<ActionResult> GetPendingPriceChange(Guid apprenticeshipKey)
         {
 	        var response = await _apiClient.Get<GetPendingPriceChangeApiResponse>(new GetPendingPriceChangeRequest(apprenticeshipKey));
-	        return Ok(new GetPendingPriceChangeResponse(response));
+
+            if(response == null || response.PendingPriceChange == null)
+            {
+                _logger.LogWarning($"No pending price change found for apprenticeship {apprenticeshipKey}");
+                return NotFound();
+            }
+
+            var ukprn = response.PendingPriceChange.Ukprn.GetValueOrDefault();
+            var providerResponse = await _apiCommitmentsClient.Get<GetProviderResponse>(new GetProviderRequest(ukprn));
+
+            if (providerResponse == null || string.IsNullOrEmpty(providerResponse.Name))
+            {
+                _logger.LogWarning($"No provider found for ukprn {ukprn}");
+                return NotFound();
+            }
+
+            return Ok(new GetPendingPriceChangeResponse(response, providerResponse.Name));
         }
 
         [HttpDelete]
@@ -97,6 +120,22 @@ namespace SFA.DAS.Apprenticeships.Api.Controllers
         public async Task<ActionResult> CancelPendingPriceChange(Guid apprenticeshipKey)
         {
             await _apiClient.Delete(new CancelPendingPriceChangeRequest(apprenticeshipKey));
+            return Ok();
+        }
+
+        [HttpPatch]
+        [Route("{apprenticeshipKey}/priceHistory/pending/reject")]
+        public async Task<ActionResult> RejectPendingPriceChange(Guid apprenticeshipKey, [FromBody] RejectPriceChangeRequest request)
+        {
+            await _apiClient.Patch(new PatchRejectApprenticeshipPriceChangeRequest(apprenticeshipKey, request.Reason));
+            return Ok();
+        }
+
+        [HttpPatch]
+        [Route("{apprenticeshipKey}/priceHistory/pending/approve")]
+        public async Task<ActionResult> ApprovePendingPriceChange(Guid apprenticeshipKey, [FromBody] ApprovePriceChangeRequest request)
+        {
+            await _apiClient.Patch(new PatchApproveApprenticeshipPriceChangeRequest(apprenticeshipKey, request.UserId));
             return Ok();
         }
     }
