@@ -1,8 +1,9 @@
 ﻿namespace SFA.DAS.Apprenticeships.Api.Models
 {
-    public class GetProviderResponse
-    {
-        public string ProviderId { get; set; }
-        public string Name { get; set; }
-    }
+public class GetProviderResponse
+{
+    public long ProviderId { get; set; }
+    public string Name { get; set; }
+}
+
 }
