@@ -3,7 +3,7 @@ using SFA.DAS.EarlyConnect.InnerApi.Requests;
 
 namespace SFA.DAS.EarlyConnect.Application.Commands.CreateStudentData
 {
-    public class CreateStudentDataCommand : IRequest<Unit>
+    public class CreateStudentDataCommand : IRequest<CreateStudentDataCommandResult>
     {
         public string RequestIP { get; set; }
         public string Payload { get; set; }
