@@ -6,7 +6,7 @@ namespace SFA.DAS.EmployerAccounts.Api.Models
     {
         public bool ShowLevyDeclarationTask { get; set; }
         public int NumberOfApprenticesToReview { get; set; }
-        public int NumberOfCohortsForApproval { get; set; }
+        public int NumberOfCohortsReadyToReview { get; set; }
         public int NumberOfPendingTransferConnections { get; set; }
         public int NumberOfTransferRequestToReview { get; set; }
         public int NumberTransferPledgeApplicationsToReview { get; set; }
@@ -17,7 +17,7 @@ namespace SFA.DAS.EmployerAccounts.Api.Models
             {
                 ShowLevyDeclarationTask = source.ShowLevyDeclarationTask,
                 NumberOfApprenticesToReview = source.NumberOfApprenticesToReview,
-                NumberOfCohortsForApproval = source.NumberOfCohortsForApproval,
+                NumberOfCohortsReadyToReview = source.NumberOfCohortsReadyToReview,
                 NumberOfPendingTransferConnections = source.NumberOfPendingTransferConnections,
                 NumberOfTransferRequestToReview = source.NumberOfTransferRequestToReview,
                 NumberTransferPledgeApplicationsToReview = source.NumberTransferPledgeApplicationsToReview
