@@ -1,9 +1,6 @@
-﻿using System.Threading.Tasks;
-using MediatR;
+﻿using MediatR;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Logging;
-using SFA.DAS.FindAnApprenticeship.Api.Models.Applications;
-using SFA.DAS.FindAnApprenticeship.Application.Queries.Apply.Index;
 
 namespace SFA.DAS.FindAnApprenticeship.Api.Controllers
 {
@@ -18,15 +15,6 @@ namespace SFA.DAS.FindAnApprenticeship.Api.Controllers
         {
             _mediator = mediator;
             _logger = logger;
-        }
-
-        [HttpGet]
-        public async Task<IActionResult> Index(string vacancyReference, [FromQuery] string applicantEmailAddress)
-        {
-            var result = await _mediator.Send(new GetIndexQuery
-                { ApplicantEmailAddress = applicantEmailAddress, VacancyReference = vacancyReference });
-
-            return Ok((GetIndexApiResponse) result);
         }
     }
 }
