@@ -6,7 +6,6 @@ using SFA.DAS.FindAnApprenticeship.InnerApi.CandidateApi.Responses;
 using SFA.DAS.FindAnApprenticeship.InnerApi.Requests;
 using SFA.DAS.FindAnApprenticeship.InnerApi.Responses;
 using SFA.DAS.SharedOuterApi.Configuration;
-using SFA.DAS.SharedOuterApi.Extensions;
 using SFA.DAS.SharedOuterApi.Interfaces;
 
 namespace SFA.DAS.FindAnApprenticeship.Application.Queries.Apply.Index;
@@ -24,53 +23,45 @@ public class GetIndexQueryHandler : IRequestHandler<GetIndexQuery,GetIndexQueryR
 
     public async Task<GetIndexQueryResult> Handle(GetIndexQuery request, CancellationToken cancellationToken)
     {
-        var result = await _findApprenticeshipApiClient.Get<GetApprenticeshipVacancyItemResponse>(new GetVacancyRequest(request.VacancyReference));
+        var application = await _candidateApiClient.Get<GetApplicationApiResponse>(new GetApplicationApiRequest(request.CandidateId, request.ApplicationId));
+        if (application == null) return null;
 
-        var putData = new PutApplicationApiRequest.PutApplicationApiRequestData
-        {
-            Email = request.ApplicantEmailAddress
-        };
-        var putRequest = new PutApplicationApiRequest(request.VacancyReference, putData);
-
-        var applicationResult = await _candidateApiClient.PutWithResponseCode<PutApplicationApiResponse>(putRequest);
-
-        applicationResult.EnsureSuccessStatusCode();
-
-        if (result is null) return null;
-        if(applicationResult is null) return null;
+        var vacancy = await _findApprenticeshipApiClient.Get<GetApprenticeshipVacancyItemResponse>(new GetVacancyRequest(application.VacancyReference));
+        if(vacancy == null) return null;
 
         return new GetIndexQueryResult
         {
-            VacancyTitle = result.Title,
-            EmployerName = result.EmployerName,
-            ClosingDate = result.ClosingDate,
-            IsDisabilityConfident = result.IsDisabilityConfident,
+            VacancyReference = vacancy.VacancyReference,
+            VacancyTitle = vacancy.Title,
+            EmployerName = vacancy.EmployerName,
+            ClosingDate = vacancy.ClosingDate,
+            IsDisabilityConfident = vacancy.IsDisabilityConfident,
             EducationHistory = new GetIndexQueryResult.EducationHistorySection
             {
-                Qualifications = applicationResult.Body.QualificationStatus,
-                TrainingCourses = applicationResult.Body.TrainingCourseStatus,
+                Qualifications = application.QualificationsStatus,
+                TrainingCourses = application.TrainingCoursesStatus,
             },
             WorkHistory = new GetIndexQueryResult.WorkHistorySection
             {
-                Jobs = applicationResult.Body.JobStatus,
-                VolunteeringAndWorkExperience = applicationResult.Body.WorkExperienceStatus,
+                Jobs = application.JobsStatus,
+                VolunteeringAndWorkExperience = application.WorkExperienceStatus,
             },
             ApplicationQuestions = new GetIndexQueryResult.ApplicationQuestionsSection
             {
-                SkillsAndStrengths = applicationResult.Body.SkillsAndStrengthsStatus,
-                WhatInterestsYou = applicationResult.Body.InterestsStatus,
-                AdditionalQuestion1 = applicationResult.Body.AdditionalQuestion1Status,
-                AdditionalQuestion1Label = result.AdditionalQuestion1,
-                AdditionalQuestion2 = applicationResult.Body.AdditionalQuestion2Status,
-                AdditionalQuestion2Label = result.AdditionalQuestion2
+                SkillsAndStrengths = application.SkillsAndStrengthStatus,
+                WhatInterestsYou = application.InterestsStatus,
+                AdditionalQuestion1 = application.AdditionalQuestion1Status,
+                AdditionalQuestion1Label = vacancy.AdditionalQuestion1,
+                AdditionalQuestion2 = application.AdditionalQuestion2Status,
+                AdditionalQuestion2Label = vacancy.AdditionalQuestion2
             },
             InterviewAdjustments = new GetIndexQueryResult.InterviewAdjustmentsSection
             {
-                RequestAdjustments = applicationResult.Body.InterviewAdjustmentsStatus
+                RequestAdjustments = application.InterviewAdjustmentsStatus
             },
             DisabilityConfidence = new GetIndexQueryResult.DisabilityConfidenceSection
             {
-                InterviewUnderDisabilityConfident = applicationResult.Body.DisabilityConfidenceStatus,
+                InterviewUnderDisabilityConfident = application.DisabilityConfidenceStatus,
             }
         };
     }
