@@ -8,15 +8,14 @@ namespace SFA.DAS.FindAnApprenticeship.InnerApi.CandidateApi.Responses
         public Guid CandidateId { get; set; }
 
         public string VacancyReference { get; set; }
-        public string Status { get; set; }
 
-        public string QualificationStatus { get; set; }
-        public string TrainingCourseStatus { get; set; }
+        public string QualificationsStatus { get; set; }
+        public string TrainingCoursesStatus { get; set; }
 
-        public string JobStatus { get; set; }
+        public string JobsStatus { get; set; }
         public string WorkExperienceStatus { get; set; }
 
-        public string SkillsAndStrengthsStatus { get; set; }
+        public string SkillsAndStrengthStatus { get; set; }
         public string InterestsStatus { get; set; }
         public string AdditionalQuestion1Status { get; set; }
         public string AdditionalQuestion2Status { get; set; }
