@@ -11,38 +11,11 @@ namespace SFA.DAS.FindApprenticeshipTraining.Api.UnitTests.Models
     public class WhenCastingToProviderBaseTypesFromMediatorType
     {
         [Test, AutoData]
-        public void Then_Maps_Fields_Appropriately_Matching_AchievementRate(
-            string sectorSubjectArea,
-            GetProvidersListItem source, GetAchievementRateItem item, GetAchievementRateItem item2)
+        public void Then_Maps_Fields_Appropriately_Returning_Null_For_AchievementRate_Data_If_No_Matching_No_AchievementRate(GetProvidersListItem source)
         {
-            item.SectorSubjectArea = sectorSubjectArea;
-            item.Level = "Two";
-            source.AchievementRates = new List<GetAchievementRateItem>
-            {
-               item,
-               item2
-            };
+            source.AchievementRates = Enumerable.Empty<GetAchievementRateItem>();
 
-            var response = new GetTrainingCourseProviderListItem().Map(source, sectorSubjectArea, 2, null, null, null, true);
-
-            response.Name.Should().Be(source.Name);
-            response.ProviderId.Should().Be(source.Ukprn);
-            response.OverallCohort.Should().Be(item.OverallCohort);
-            response.HasLocation.Should().BeTrue();
-            response.OverallAchievementRate.Should().Be(item.OverallAchievementRate);
-
-        }
-        [Test, AutoData]
-        public void Then_Maps_Fields_Appropriately_Returning_Null_For_AchievementRate_Data_If_No_Matching_No_AchievementRate(string sectorSubjectArea,
-            GetProvidersListItem source, GetAchievementRateItem item, GetAchievementRateItem item2)
-        {
-            source.AchievementRates = new List<GetAchievementRateItem>
-            {
-                item,
-                item2
-            };
-
-            var response = new GetTrainingCourseProviderListItem().Map(source, sectorSubjectArea, 1, null, null, null, true);
+            var response = new GetTrainingCourseProviderListItem().Map(source, null, null, null, true);
 
             response.Name.Should().Be(source.Name);
             response.ProviderId.Should().Be(source.Ukprn);
@@ -52,12 +25,12 @@ namespace SFA.DAS.FindApprenticeshipTraining.Api.UnitTests.Models
         }
 
         [Test, AutoData]
-        public void Then_Maps_Fields_Appropriately_Returning_Null_For_AchievementRate_Data_If_No_AchievementRates_And_Empty_List_For_DeliveryModes_If_No_Delivery_Modes(string sectorSubjectArea, GetProvidersListItem source)
+        public void Then_Maps_Fields_Appropriately_Returning_Null_For_AchievementRate_Data_If_No_AchievementRates_And_Empty_List_For_DeliveryModes_If_No_Delivery_Modes(GetProvidersListItem source)
         {
-            source.AchievementRates = null;
+            source.AchievementRates = Enumerable.Empty<GetAchievementRateItem>();
             source.DeliveryModels = new List<DeliveryModel>();
 
-            var response = new GetTrainingCourseProviderListItem().Map(source, sectorSubjectArea, 1, new List<DeliveryModeType>(), new List<FeedbackRatingType>(), new List<FeedbackRatingType>(), true);
+            var response = new GetTrainingCourseProviderListItem().Map(source, new List<DeliveryModeType>(), new List<FeedbackRatingType>(), new List<FeedbackRatingType>(), true);
 
             response.Name.Should().Be(source.Name);
             response.ProviderId.Should().Be(source.Ukprn);
@@ -69,9 +42,8 @@ namespace SFA.DAS.FindApprenticeshipTraining.Api.UnitTests.Models
         }
 
         [Test, AutoData]
-        public void Then_Maps_Delivery_Types_Returning_The_Smallest_Distance_Only(string sectorSubjectArea, GetProvidersListItem source)
+        public void Then_Maps_Delivery_Types_Returning_The_Smallest_Distance_Only(GetProvidersListItem source)
         {
-            source.AchievementRates = null;
             source.DeliveryModels = new List<DeliveryModel>
             {
                 new()
@@ -97,7 +69,7 @@ namespace SFA.DAS.FindApprenticeshipTraining.Api.UnitTests.Models
                 }
             };
 
-            var response = new GetTrainingCourseProviderListItem().Map(source, sectorSubjectArea, 1, new List<DeliveryModeType>(), new List<FeedbackRatingType>(), new List<FeedbackRatingType>(), true);
+            var response = new GetTrainingCourseProviderListItem().Map(source, new List<DeliveryModeType>(), new List<FeedbackRatingType>(), new List<FeedbackRatingType>(), true);
 
             response.DeliveryModes.Count.Should().Be(3);
             response.DeliveryModes.FirstOrDefault(c => c.DeliveryModeType == DeliveryModeType.BlockRelease)?.DistanceInMiles.Should().Be(3.1m);
@@ -109,9 +81,8 @@ namespace SFA.DAS.FindApprenticeshipTraining.Api.UnitTests.Models
         [Test]
         [InlineAutoData("BlockRelease", DeliveryModeType.BlockRelease)]
         [InlineAutoData("DayRelease", DeliveryModeType.DayRelease)]
-        public void Then_Maps_Delivery_Types_Returning_The_Smallest_Distance_Only_For_One_Type(string deliveryModeString, DeliveryModeType deliveryModeType, string sectorSubjectArea, GetProvidersListItem source)
+        public void Then_Maps_Delivery_Types_Returning_The_Smallest_Distance_Only_For_One_Type(string deliveryModeString, DeliveryModeType deliveryModeType, GetProvidersListItem source)
         {
-            source.AchievementRates = null;
             source.DeliveryModels = new List<DeliveryModel>
             {
                 new()
@@ -134,7 +105,7 @@ namespace SFA.DAS.FindApprenticeshipTraining.Api.UnitTests.Models
                 }
             };
 
-            var response = new GetTrainingCourseProviderListItem().Map(source, sectorSubjectArea, 1, new List<DeliveryModeType>(), new List<FeedbackRatingType>(), new List<FeedbackRatingType>(), true);
+            var response = new GetTrainingCourseProviderListItem().Map(source, new List<DeliveryModeType>(), new List<FeedbackRatingType>(), new List<FeedbackRatingType>(), true);
 
             response.DeliveryModes.Count.Should().Be(1);
             response.DeliveryModes.FirstOrDefault(c => c.DeliveryModeType == deliveryModeType)?.DistanceInMiles.Should().Be(2.5m);
@@ -142,29 +113,28 @@ namespace SFA.DAS.FindApprenticeshipTraining.Api.UnitTests.Models
         }
 
         [Test, AutoData]
-        public void Then_Maps_All_DeliveryType_Fields_And_Sets_At_WorkPlace_Distance_To_Zero(string sectorSubjectArea, GetProvidersListItem source,DeliveryModel item) 
+        public void Then_Maps_All_DeliveryType_Fields_And_Sets_At_WorkPlace_Distance_To_Zero(GetProvidersListItem source, DeliveryModel item)
         {
-            source.AchievementRates = null;
             item.LocationType = LocationType.Regional;
-            source.DeliveryModels = new List<DeliveryModel> {item};
-            var response = new GetTrainingCourseProviderListItem().Map(source, sectorSubjectArea, 1, new List<DeliveryModeType>(), new List<FeedbackRatingType>(), new List<FeedbackRatingType>(), true);
+            source.DeliveryModels = new List<DeliveryModel> { item };
+            var response = new GetTrainingCourseProviderListItem().Map(source, new List<DeliveryModeType>(), new List<FeedbackRatingType>(), new List<FeedbackRatingType>(), true);
 
             response.DeliveryModes.First().Should().BeEquivalentTo(item, options => options
-                .Excluding(c=>c.LocationType)
+                .Excluding(c => c.LocationType)
                 .Excluding(c => c.DayRelease)
-                .Excluding(c=>c.BlockRelease)
+                .Excluding(c => c.BlockRelease)
                 .Excluding(c => c.DistanceInMiles));
             response.DeliveryModes.First().DeliveryModeType.Should().Be(DeliveryModeType.Workplace);
             response.DeliveryModes.First().DistanceInMiles.Should().Be(0m);
         }
 
         [Test, AutoData]
-        public void Then_Maps_Zero_If_No_Feedback(GetProvidersListItem source, string sectorSubjectArea)
+        public void Then_Maps_Zero_If_No_Feedback(GetProvidersListItem source)
         {
             source.EmployerFeedback.ReviewCount = 0;
             source.ApprenticeFeedback.ReviewCount = 0;
-            
-            var response = new GetTrainingCourseProviderListItem().Map(source, sectorSubjectArea, 1, new List<DeliveryModeType>(), new List<FeedbackRatingType>(), new List<FeedbackRatingType>(), true);
+
+            var response = new GetTrainingCourseProviderListItem().Map(source, new List<DeliveryModeType>(), new List<FeedbackRatingType>(), new List<FeedbackRatingType>(), true);
 
             response.EmployerFeedback.TotalEmployerResponses.Should().Be(0);
             response.EmployerFeedback.TotalFeedbackRating.Should().Be(0);
@@ -173,14 +143,14 @@ namespace SFA.DAS.FindApprenticeshipTraining.Api.UnitTests.Models
         }
 
         [Test, AutoData]
-        public void Then_Maps_ResponsesAndRatings(GetProvidersListItem source, string sectorSubjectArea)
+        public void Then_Maps_ResponsesAndRatings(GetProvidersListItem source)
         {
             source.EmployerFeedback.ReviewCount = 34;
             source.EmployerFeedback.Stars = 3;
             source.ApprenticeFeedback.ReviewCount = 24;
             source.ApprenticeFeedback.Stars = 2;
 
-            var response = new GetTrainingCourseProviderListItem().Map(source, sectorSubjectArea, 1, new List<DeliveryModeType>(), new List<FeedbackRatingType>(), new List<FeedbackRatingType>(), true);
+            var response = new GetTrainingCourseProviderListItem().Map(source, new List<DeliveryModeType>(), new List<FeedbackRatingType>(), new List<FeedbackRatingType>(), true);
 
             response.EmployerFeedback.TotalFeedbackRating.Should().Be(3);
             response.EmployerFeedback.TotalEmployerResponses.Should().Be(34);
@@ -189,10 +159,10 @@ namespace SFA.DAS.FindApprenticeshipTraining.Api.UnitTests.Models
         }
 
         [Test, AutoData]
-        public void Then_Maps_Not_Found_Delivery_Mode(string sectorSubjectArea, GetProvidersListItem source)
+        public void Then_Maps_Not_Found_Delivery_Mode(GetProvidersListItem source)
         {
             source.DeliveryModels = new List<DeliveryModel>();
-            var response = new GetTrainingCourseProviderListItem().Map(source, sectorSubjectArea, 1, new List<DeliveryModeType>(), new List<FeedbackRatingType>(), new List<FeedbackRatingType>(), true);
+            var response = new GetTrainingCourseProviderListItem().Map(source, new List<DeliveryModeType>(), new List<FeedbackRatingType>(), new List<FeedbackRatingType>(), true);
 
             response.DeliveryModes.First().DeliveryModeType.Should().Be(DeliveryModeType.NotFound);
             response.DeliveryModes.First().Address1.Should().BeNullOrEmpty();
@@ -204,9 +174,8 @@ namespace SFA.DAS.FindApprenticeshipTraining.Api.UnitTests.Models
         }
 
         [Test, AutoData]
-        public void Then_If_Delivery_Modes_Are_Passed_The_Results_Are_Filtered(string sectorSubjectArea, GetProvidersListItem source)
+        public void Then_If_Delivery_Modes_Are_Passed_The_Results_Are_Filtered(GetProvidersListItem source)
         {
-            source.AchievementRates = null;
             source.DeliveryModels = new List<DeliveryModel>
             {
                 new()
@@ -230,7 +199,7 @@ namespace SFA.DAS.FindApprenticeshipTraining.Api.UnitTests.Models
                 }
             };
 
-            var response = new GetTrainingCourseProviderListItem().Map(source, sectorSubjectArea, 1, new List<DeliveryModeType>
+            var response = new GetTrainingCourseProviderListItem().Map(source, new List<DeliveryModeType>
             {
                 DeliveryModeType.DayRelease
             }, new List<FeedbackRatingType>(), new List<FeedbackRatingType>(), true);
@@ -240,9 +209,8 @@ namespace SFA.DAS.FindApprenticeshipTraining.Api.UnitTests.Models
         }
 
         [Test, AutoData]
-        public void Then_If_There_Are_Multiple_Delivery_Modes_Filtered_And_Not_Match_Then_It_Is_Returned_Correctly(string sectorSubjectArea, GetProvidersListItem source)
+        public void Then_If_There_Are_Multiple_Delivery_Modes_Filtered_And_Not_Match_Then_It_Is_Returned_Correctly(GetProvidersListItem source)
         {
-            source.AchievementRates = null;
             source.DeliveryTypes = new List<GetDeliveryTypeItem>
             {
                 new GetDeliveryTypeItem
@@ -252,7 +220,7 @@ namespace SFA.DAS.FindApprenticeshipTraining.Api.UnitTests.Models
                 }
             };
 
-            var response = new GetTrainingCourseProviderListItem().Map(source, sectorSubjectArea, 1, new List<DeliveryModeType>
+            var response = new GetTrainingCourseProviderListItem().Map(source, new List<DeliveryModeType>
             {
                 DeliveryModeType.DayRelease,
                 DeliveryModeType.BlockRelease
@@ -261,9 +229,8 @@ namespace SFA.DAS.FindApprenticeshipTraining.Api.UnitTests.Models
             response.Should().NotBeNull();
         }
         [Test, AutoData]
-        public void Then_If_There_Are_Multiple_Delivery_Modes_Filtered_And_Match_Then_It_Is_Returned_Correctly(string sectorSubjectArea, GetProvidersListItem source)
+        public void Then_If_There_Are_Multiple_Delivery_Modes_Filtered_And_Match_Then_It_Is_Returned_Correctly(GetProvidersListItem source)
         {
-            source.AchievementRates = null;
             source.DeliveryModels = new List<DeliveryModel>
             {
                 new()
@@ -281,7 +248,7 @@ namespace SFA.DAS.FindApprenticeshipTraining.Api.UnitTests.Models
                 }
             };
 
-            var response = new GetTrainingCourseProviderListItem().Map(source, sectorSubjectArea, 1, new List<DeliveryModeType>
+            var response = new GetTrainingCourseProviderListItem().Map(source, new List<DeliveryModeType>
             {
                 DeliveryModeType.DayRelease,
                 DeliveryModeType.Workplace
@@ -291,9 +258,8 @@ namespace SFA.DAS.FindApprenticeshipTraining.Api.UnitTests.Models
         }
 
         [Test, AutoData]
-        public void The_If_There_Are_Delivery_Modes_To_Filter_And_Returns_No_Delivery_Modes_After_Filter_Then_Null_Returned(string sectorSubjectArea, GetProvidersListItem source)
+        public void The_If_There_Are_Delivery_Modes_To_Filter_And_Returns_No_Delivery_Modes_After_Filter_Then_Null_Returned(GetProvidersListItem source)
         {
-            source.AchievementRates = null;
             source.DeliveryTypes = new List<GetDeliveryTypeItem>
             {
                 new GetDeliveryTypeItem
@@ -303,7 +269,7 @@ namespace SFA.DAS.FindApprenticeshipTraining.Api.UnitTests.Models
                 }
             };
 
-            var response = new GetTrainingCourseProviderListItem().Map(source, sectorSubjectArea, 1, new List<DeliveryModeType>
+            var response = new GetTrainingCourseProviderListItem().Map(source, new List<DeliveryModeType>
             {
                 DeliveryModeType.DayRelease
             }, new List<FeedbackRatingType>(), new List<FeedbackRatingType>(), true);
@@ -312,9 +278,8 @@ namespace SFA.DAS.FindApprenticeshipTraining.Api.UnitTests.Models
         }
 
         [Test, AutoData]
-        public void Then_If_National_At_Workplace_Is_Selected_As_Delivery_Mode_Filter_Then_Null_Returned_If_No_National_At_Workplace(string sectorSubjectArea, GetProvidersListItem source)
+        public void Then_If_National_At_Workplace_Is_Selected_As_Delivery_Mode_Filter_Then_Null_Returned_If_No_National_At_Workplace(GetProvidersListItem source)
         {
-            source.AchievementRates = null;
             source.DeliveryModels = new List<DeliveryModel>
             {
                 new()
@@ -326,7 +291,7 @@ namespace SFA.DAS.FindApprenticeshipTraining.Api.UnitTests.Models
                 }
             };
 
-            var response = new GetTrainingCourseProviderListItem().Map(source, sectorSubjectArea, 1, new List<DeliveryModeType>
+            var response = new GetTrainingCourseProviderListItem().Map(source, new List<DeliveryModeType>
             {
                 DeliveryModeType.Workplace,
                 DeliveryModeType.National
@@ -336,9 +301,8 @@ namespace SFA.DAS.FindApprenticeshipTraining.Api.UnitTests.Models
         }
 
         [Test, AutoData]
-        public void Then_If_National_At_Workplace_Is_Selected_As_Delivery_Mode_Filter_Then_Not_Null_Returned_If_National_At_Workplace(string sectorSubjectArea, GetProvidersListItem source)
+        public void Then_If_National_At_Workplace_Is_Selected_As_Delivery_Mode_Filter_Then_Not_Null_Returned_If_National_At_Workplace(GetProvidersListItem source)
         {
-            source.AchievementRates = null;
             source.DeliveryModels = new List<DeliveryModel>
             {
                 new()
@@ -350,7 +314,7 @@ namespace SFA.DAS.FindApprenticeshipTraining.Api.UnitTests.Models
                 }
             };
 
-            var response = new GetTrainingCourseProviderListItem().Map(source, sectorSubjectArea, 1, new List<DeliveryModeType>
+            var response = new GetTrainingCourseProviderListItem().Map(source, new List<DeliveryModeType>
             {
                 DeliveryModeType.Workplace,
                 DeliveryModeType.National
@@ -360,12 +324,12 @@ namespace SFA.DAS.FindApprenticeshipTraining.Api.UnitTests.Models
         }
 
         [Test, AutoData]
-        public void Then_If_There_Are_EmployerRatings_To_Filter_Then_Matches_On_Values(string sectorSubjectArea, GetProvidersListItem source)
+        public void Then_If_There_Are_EmployerRatings_To_Filter_Then_Matches_On_Values(GetProvidersListItem source)
         {
             source.EmployerFeedback.ReviewCount = 1;
             source.EmployerFeedback.Stars = 4;
 
-            var response = new GetTrainingCourseProviderListItem().Map(source, sectorSubjectArea, 1, new List<DeliveryModeType>(), new List<FeedbackRatingType> { FeedbackRatingType.Poor }, new List<FeedbackRatingType>(), true);
+            var response = new GetTrainingCourseProviderListItem().Map(source, new List<DeliveryModeType>(), new List<FeedbackRatingType> { FeedbackRatingType.Poor }, new List<FeedbackRatingType>(), true);
 
             response.Should().BeNull();
 
@@ -373,12 +337,12 @@ namespace SFA.DAS.FindApprenticeshipTraining.Api.UnitTests.Models
 
 
         [Test, AutoData]
-        public void Then_If_There_Are_Multiple_EmployerRatings_To_Filter_Then_Matches_On_Values(string sectorSubjectArea, GetProvidersListItem source)
+        public void Then_If_There_Are_Multiple_EmployerRatings_To_Filter_Then_Matches_On_Values(GetProvidersListItem source)
         {
             source.EmployerFeedback.ReviewCount = 3;
             source.EmployerFeedback.Stars = 4;
 
-            var response = new GetTrainingCourseProviderListItem().Map(source, sectorSubjectArea, 1, new List<DeliveryModeType>(), new List<FeedbackRatingType> { FeedbackRatingType.Poor, FeedbackRatingType.Excellent }, new List<FeedbackRatingType>(), true);
+            var response = new GetTrainingCourseProviderListItem().Map(source, new List<DeliveryModeType>(), new List<FeedbackRatingType> { FeedbackRatingType.Poor, FeedbackRatingType.Excellent }, new List<FeedbackRatingType>(), true);
 
             response.Should().NotBeNull();
             response.Should().BeEquivalentTo(source, options => options.ExcludingMissingMembers());
@@ -386,12 +350,12 @@ namespace SFA.DAS.FindApprenticeshipTraining.Api.UnitTests.Models
         }
 
         [Test, AutoData]
-        public void Then_If_There_Are_ApprenticeRatings_To_Filter_Then_Matches_On_Values(string sectorSubjectArea, GetProvidersListItem source)
+        public void Then_If_There_Are_ApprenticeRatings_To_Filter_Then_Matches_On_Values(GetProvidersListItem source)
         {
             source.ApprenticeFeedback.ReviewCount = 23;
             source.ApprenticeFeedback.Stars = 4;
 
-            var response = new GetTrainingCourseProviderListItem().Map(source, sectorSubjectArea, 1, new List<DeliveryModeType>(), new List<FeedbackRatingType> { FeedbackRatingType.Poor }, new List<FeedbackRatingType>(), true);
+            var response = new GetTrainingCourseProviderListItem().Map(source, new List<DeliveryModeType>(), new List<FeedbackRatingType> { FeedbackRatingType.Poor }, new List<FeedbackRatingType>(), true);
 
             response.Should().BeNull();
 
@@ -399,10 +363,10 @@ namespace SFA.DAS.FindApprenticeshipTraining.Api.UnitTests.Models
 
 
         [Test, AutoData]
-        public void Then_If_There_Are_Multiple_ApprenticeRatings_To_Filter_Then_Matches_On_Values(string sectorSubjectArea, GetProvidersListItem source)
+        public void Then_If_There_Are_Multiple_ApprenticeRatings_To_Filter_Then_Matches_On_Values(GetProvidersListItem source)
         {
             source.ApprenticeFeedback.Stars = 4;
-            var response = new GetTrainingCourseProviderListItem().Map(source, sectorSubjectArea, 1, new List<DeliveryModeType>(), new List<FeedbackRatingType>(), new List<FeedbackRatingType> { FeedbackRatingType.Poor, FeedbackRatingType.Excellent }, true);
+            var response = new GetTrainingCourseProviderListItem().Map(source, new List<DeliveryModeType>(), new List<FeedbackRatingType>(), new List<FeedbackRatingType> { FeedbackRatingType.Poor, FeedbackRatingType.Excellent }, true);
 
             response.Should().NotBeNull();
             response.Should().BeEquivalentTo(source, options => options.ExcludingMissingMembers());
