@@ -1,6 +1,11 @@
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Net;
+using System.Threading;
+using System.Threading.Tasks;
 using AutoFixture.NUnit3;
 using FluentAssertions;
-using Microsoft.Extensions.Azure;
 using Moq;
 using NUnit.Framework;
 using SFA.DAS.FindApprenticeshipTraining.Application.TrainingCourses.Queries.GetTrainingCourseProvider;
@@ -13,13 +18,6 @@ using SFA.DAS.SharedOuterApi.InnerApi.Requests;
 using SFA.DAS.SharedOuterApi.Interfaces;
 using SFA.DAS.SharedOuterApi.Models;
 using SFA.DAS.Testing.AutoFixture;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Net;
-using System.Threading;
-using System.Threading.Tasks;
-using SFA.DAS.FindApprenticeshipTraining.Services;
 
 namespace SFA.DAS.FindApprenticeshipTraining.UnitTests.Application.TrainingCourses.Queries
 {
@@ -147,7 +145,7 @@ namespace SFA.DAS.FindApprenticeshipTraining.UnitTests.Application.TrainingCours
         {
             mockLookupService
                 .Setup(l => l.GetLocationInformation(query.Location, query.Lat, query.Lon, false))
-                .ReturnsAsync(new LocationItem (query.Location, new[] { query.Lat, query.Lon }, string.Empty));
+                .ReturnsAsync(new LocationItem(query.Location, new[] { query.Lat, query.Lon }, string.Empty));
             var matchingItem = new ShortlistItem
             {
                 Id = Guid.NewGuid(),
@@ -257,7 +255,7 @@ namespace SFA.DAS.FindApprenticeshipTraining.UnitTests.Application.TrainingCours
                 .Setup(client => client.Get<GetStandardsListItem>(It.Is<GetStandardRequest>(c => c.GetUrl.Contains(query.CourseId.ToString()))))
                 .ReturnsAsync(apiCourseResponse);
             mockRoatpV2ApiClient.Setup(client => client.Get<GetOverallAchievementRateResponse>(It.Is<GetOverallAchievementRateRequest>(c =>
-                    c.GetUrl.Contains(apiCourseResponse.SectorSubjectAreaTier2Description)
+                    c.GetUrl.Contains(apiCourseResponse.SectorSubjectAreaTier1.ToString())
                 )))
                 .ReturnsAsync(apiAchievementRateResponse);
             mockRoatpV2ApiClient
@@ -362,7 +360,7 @@ namespace SFA.DAS.FindApprenticeshipTraining.UnitTests.Application.TrainingCours
 
             var result = await handler.Handle(query, CancellationToken.None);
 
-            result.AdditionalCourses.Should().BeEquivalentTo(filteredProvidersStandardResponse, options => 
+            result.AdditionalCourses.Should().BeEquivalentTo(filteredProvidersStandardResponse, options =>
                 options
                     .Excluding(c => c.IsApprovedByRegulator)
                     .Excluding(c => c.ApprovalBody)
@@ -484,7 +482,7 @@ namespace SFA.DAS.FindApprenticeshipTraining.UnitTests.Application.TrainingCours
                 .ReturnsAsync(allCoursesApiResponse);
             mockRoatpV2ApiClient.Setup(client => client.Get<GetOverallAchievementRateResponse>(It.Is<GetOverallAchievementRateRequest>(
                     c =>
-                        c.GetUrl.Contains(apiCourseResponse.SectorSubjectAreaTier2Description)
+                        c.GetUrl.Contains(apiCourseResponse.SectorSubjectAreaTier1.ToString())
                 )))
                 .ReturnsAsync(apiAchievementRateResponse);
             apprenticeFeedbackResponse.Ukprn = query.ProviderId;
@@ -543,7 +541,7 @@ namespace SFA.DAS.FindApprenticeshipTraining.UnitTests.Application.TrainingCours
 
             mockRoatpV2ApiClient.Setup(client => client.Get<GetOverallAchievementRateResponse>(It.Is<GetOverallAchievementRateRequest>(
                     c =>
-                        c.GetUrl.Contains(apiCourseResponse.SectorSubjectAreaTier2Description)
+                        c.GetUrl.Contains(apiCourseResponse.SectorSubjectAreaTier1.ToString())
                 )))
                 .ReturnsAsync(apiAchievementRateResponse);
 
