@@ -15,7 +15,7 @@ namespace SFA.DAS.FindApprenticeshipTraining.InnerApi.Responses
         public string Phone { get; set; }
         public int StandardId { get; set; }
         public Guid? ShortlistId { get; set; }
-        public IEnumerable<GetAchievementRateItem> AchievementRates { get; set; }
+        public IEnumerable<GetAchievementRateItem> AchievementRates { get; set; } = Enumerable.Empty<GetAchievementRateItem>();
         public IEnumerable<GetDeliveryTypeItem> DeliveryTypes { get; set; }
         public IEnumerable<DeliveryModel> DeliveryModels { get; set; }
         public GetEmployerFeedbackResponse EmployerFeedback { get; set; }
