@@ -51,9 +51,9 @@ namespace SFA.DAS.SharedOuterApi.Services
             throw new System.NotImplementedException();
         }
 
-        public Task Delete(IDeleteApiRequest request)
+        public async Task Delete(IDeleteApiRequest request)
         {
-            throw new System.NotImplementedException();
+            await _apiClient.Delete(request);
         }
 
         public Task Patch<TData>(IPatchApiRequest<TData> request)
@@ -78,7 +78,7 @@ namespace SFA.DAS.SharedOuterApi.Services
 
         public Task<ApiResponse<string>> PatchWithResponseCode<TData>(IPatchApiRequest<TData> request)
         {
-            throw new System.NotImplementedException();
+            return _apiClient.PatchWithResponseCode<TData>(request);
         }
         public Task<ApiResponse<TResponse>> PutWithResponseCode<TResponse>(IPutApiRequest request)
         {
